+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* Add `PlaceAwareActivity` to make it easy for activities to gain access to place that activated activity.
+
 ## 0.12
 * Bump the version of gwt tested against to `2.8.2-v20191108`
 * Convert gwt dependency into a "provided" dependency.